+++ src/components/Callout/Callout.tsx
@@ -1,7 +1,7 @@
 import type { FC } from 'react';
-import type { CalloutProps } from './Callout.types';
+import type { CalloutContainerProps, CalloutProps } from './Callout.types';
 
-import styled from 'styled-components';
+import styled, { css } from 'styled-components';
 import { forwardRef } from 'react';
 
 import { ColorTypes, SpaceSizes } from '../../theme';
@@ -11,32 +11,69 @@ import { Text } from '../typographyLegacy';
 import { Icon } from '../Icon';
 import { TextSizes } from '../typographyLegacy/Text/Text.enums';
 import { SSCIconNames } from '../../theme/icons/icons.enums';
+import { CalloutColors } from './Callout.enums';
 import { CLX_COMPONENT } from '../../theme/constants';
 
-const IconContainer = styled.div`
+const CalloutIconNeutral = css`
+  background-color: ${getColor(ColorTypes.neutral300)};
+`;
+const CalloutIconInfo = css`
+  background-color: ${getColor(ColorTypes.info100)};
+`;
+const calloutIconColors = {
+  [CalloutColors.neutral]: CalloutIconNeutral,
+  [CalloutColors.info]: CalloutIconInfo,
+};
+
+const IconContainer = styled.div<CalloutContainerProps>`
   width: ${pxToRem(36)};
   height: ${pxToRem(36)};
   flex-shrink: 0;
-  background-color: ${getColor(ColorTypes.info700)};
   border-radius: ${getRadii('circle')};
   display: flex;
   align-items: center;
   justify-content: center;
   font-size: ${getFontSize('mdPlus')};
+  ${({ $color }) => calloutIconColors[$color]};
 `;
 
-const Container = styled(Padbox)`
+const CalloutNeutral = css`
+  background-color: ${getColor(ColorTypes.neutral100)};
+`;
+const CalloutInfo = css`
   background-color: ${getColor(ColorTypes.info50)};
+`;
+
+const calloutColors = {
+  [CalloutColors.neutral]: CalloutNeutral,
+  [CalloutColors.info]: CalloutInfo,
+};
+
+const Container = styled(Padbox)<CalloutContainerProps>`
   border-radius: ${getRadii('default')};
+  ${({ $color }) => calloutColors[$color]};
 `;
 
 const Callout: FC<CalloutProps> = forwardRef<HTMLSpanElement, CalloutProps>(
-  ({ children, icon = SSCIconNames.lightbulb }, ref) => (
-    <Container ref={ref} className={CLX_COMPONENT} paddingSize={SpaceSizes.md}>
+  (
+    { children, icon = SSCIconNames.lightbulb, color = CalloutColors.info },
+    ref,
+  ) => (
+    <Container
+      ref={ref}
+      $color={color}
+      className={CLX_COMPONENT}
+      paddingSize={SpaceSizes.md}
+    >
       <Inline gap={SpaceSizes.md}>
-        <IconContainer>
+        <IconContainer $color={color}>
           {typeof icon === 'string' ? (
-            <Icon color={ColorTypes.neutral0} name={icon} />
+            <Icon
+              color={
+                color === 'info' ? ColorTypes.info700 : ColorTypes.neutral700
+              }
+              name={icon}
+            />
           ) : (
             icon
           )}