+++ src/managers/NotificationsManager/api.ts
@@ -14,6 +14,12 @@ export const removeNotification = (id: string) => {
   );
 };
 
+const setupAutodismiss = (notification: Notification) => {
+  if (notification.autoDismiss || notification.autoDismiss === undefined) {
+    setTimeout(() => removeNotification(notification.id), 6000);
+  }
+};
+
 export const addNotification = (notification: Notification) => {
   window.dispatchEvent(
     createCustomEvent<typeof ACTIONS, { notification: Notification }>(
@@ -23,7 +29,17 @@ export const addNotification = (notification: Notification) => {
       },
     ),
   );
-  if (notification.autoDismiss || notification.autoDismiss === undefined) {
-    setTimeout(() => removeNotification(notification.id), 6000);
-  }
+  setupAutodismiss(notification);
+};
+
+export const updateNotification = (notification: Notification) => {
+  window.dispatchEvent(
+    createCustomEvent<typeof ACTIONS, { notification: Notification }>(
+      ACTIONS.UPDATE_NOTIFICATION,
+      {
+        notification,
+      },
+    ),
+  );
+  setupAutodismiss(notification);
 };