+++ src/theme/radii.enums.ts
@@ -1,5 +1,6 @@
 export const RadiusTypes = {
   default: 'default',
+  double: 'double',
   half: 'half',
   circle: 'circle',
   round: 'round',